Bug fixes to get DOCN+FV3 to run.

+++ scripts/exhafs_forecast.sh
@@ -500,21 +500,12 @@ if [ ${run_datm} = yes ];  then
       nems.configure.cdeps.tmp > nems.configure
 
 elif [ ${run_docn} = yes ];  then
-
-  now=${CDATE:0:8}00
-  end=$( date -d "${CDATE:0:4}-${CDATE:4:2}-${CDATE:6:2}t${CDATE:8:2}:00:00+00 +$NHRS hours" +%Y%m%d )00
-  while (( now <= end )) ; do
-    infile="$DOCNdir/oisst-avhrr-v02r01.${now:0:8}.nc"
-    linkfile="INPUT/oisst-avhrr-v02r01.${now:0:8}.nc"
-    if [[ ! -s "$infile" ]] ; then
-        echo "Error: $infile does not exist or is zero size." 1>&2
-        exit 1
-    fi
-    ln -sf "$infile" "$linkfile"
-    now=$( "$NDATE" +24 $now )
-  done
-
-#  ln -sf $DOCNdir/*.nc INPUT/
+  MAKE_MESH_OCN=$( echo "${make_mesh_ocn:-no}" | tr a-z A-Z )
+  if [[ "$MAKE_MESH_OCN" == YES ]] ; then
+      ln -sf "$merged_docn_input" INPUT/
+  else
+      ln -sf "$DOCNdir"/*.nc INPUT/
+  fi
 
   cp ${PARMhafs}/cdeps/docn_in .
   cp ${PARMhafs}/cdeps/docn.streams.xml .
@@ -526,11 +517,13 @@ elif [ ${run_docn} = yes ];  then
   sed -i "s/_yearLast_/$endyr/g" docn.streams.xml
 
   sed -i "s/_mesh_ocn_/INPUT\/$(basename $mesh_ocn)/g" docn.streams.xml
-  for file in `ls INPUT/sst*.nc ` ; do
-    sed -i "/<\/stream_data_files>/i \ \ \ \ \ \ <file>$file<\/file>" docn.streams.xml
+  for file in INPUT/oisst*.nc INPUT/sst*.nc ; do
+    if [[ -s "$file" ]] ; then
+      sed -i "/<\/stream_data_files>/i \ \ \ \ \ \ <file>$file<\/file>" docn.streams.xml
+    fi
   done
 
-  ln -sf ${mesh_ocn} INPUT/
+  ln -sf "${mesh_ocn}" INPUT/
 
   cp ${PARMforecast}/nems.configure.cdeps.tmp ./
   sed -e "s/_ATM_petlist_bounds_/${ATM_petlist_bounds}/g" \
